feat(cli-mcp): add mcp command and apply cli - change folder structure(index,server)
