+++ packages/cli/src/index.ts
@@ -4,6 +4,7 @@ import { Command } from "commander"
 
 import { addCommand } from "@/commands/add"
 import { initCommand } from "@/commands/init"
+import { mcpCommand } from "@/commands/mcp"
 import { codemodCommand } from "@/commands/radix-import"
 
 import packageJson from "../package.json"
@@ -16,6 +17,7 @@ async function main() {
     .addCommand(initCommand)
     .addCommand(addCommand)
     .addCommand(codemodCommand)
+    .addCommand(mcpCommand)
     .parse()
 }
 main()
